% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btm_annotate.R
\name{btm_annotate}
\alias{btm_annotate}
\title{Annotate text}
\usage{
btm_annotate(data, modeltype = c("NAV", "NPN", "ADJ"))
}
\arguments{
\item{data}{Data frame; the column to be annotated must be named \code{text}}

\item{modeltype}{Type of BTM model that we want to run (noun, adjective, verb; noun, proper noun; adjective only)}
}
\value{
A list with original data, annotated text, biterm co-occurrence counts by document, and a subset of annotated data based upon model type
}
\description{
Using the \href{https://cran.r-project.org/web/packages/udpipe/index.html}{udpipe} pacakge to annotate text
}
\examples{
\dontrun{
btm_annotate(text_data, "NAV")
}
}
