% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btm_best_topic.R
\name{btm_best_topic}
\alias{btm_best_topic}
\title{Find the best topic}
\usage{
btm_best_topic(pca, coherence_df)
}
\arguments{
\item{pca}{PCA coordinates from \code{btm_pca_coords()} function}

\item{coherence_df}{Coherence data from \code{btm_coherence()} function}
}
\value{
List of fulldata and max and min coordinates for plotting
}
\description{
Identify the best topic via the lowest coherence score
}
\examples{
\dontrun{
best_topic <- btm_best_topic(pca_coords, coherence)
}
}
