% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btm_buildDTM.R
\name{btm_buildDTM}
\alias{btm_buildDTM}
\title{Build DTM}
\usage{
btm_buildDTM(data)
}
\arguments{
\item{data}{List of data from \code{btm_annotate()} function}
}
\value{
DocumentTermMatrix object for use with topicmodels package
}
\description{
Build DTM
}
\examples{
\dontrun{
dtm <- btm_buildDTM(NAV_reviews)
}
}
