% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btm_pca_coords.R
\name{btm_pca_coords}
\alias{btm_pca_coords}
\title{Get PCA Coordinates}
\usage{
btm_pca_coords(model, data)
}
\arguments{
\item{model}{Topic models}

\item{data}{Annotated data}
}
\value{
list with x,y coordinates and term data for each topic
}
\description{
Using some code from the LDAvis package to find the x,y coordinates for visualization
}
\examples{
\dontrun{
pca_coords <- purrr::map(testing_models, btm_pca_coords, data = NAV_reviews)
}
}
