% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btm_topicmodel.R
\name{btm_topicmodel}
\alias{btm_topicmodel}
\title{Topic Model with BTM}
\usage{
btm_topicmodel(n_topics, data)
}
\arguments{
\item{n_topics}{Number of topics to run}

\item{data}{List output from \code{btm_annotate()} function}
}
\value{
BTM object
}
\description{
Run the BTM topic model function over the data and biterms
}
\examples{
\dontrun{
n_topics <- 5:20
testing_models <- n_topics \%>\%
  map(btm_topicmodel, data = NAV_reviews) \%>\%
  set_names(nm = paste0("topicsize_", n_topics))
}
}
