% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete_btm.R
\name{complete_btm}
\alias{complete_btm}
\title{Complete BTM (topic modeling)}
\usage{
complete_btm(data, min_topics, max_topics, model)
}
\arguments{
\item{data}{Original data.frame/tibble with one column named "text"}

\item{min_topics}{Minimum number of topics to model}

\item{max_topics}{Maximum number of topics to model}

\item{model}{NAV (noun, adjective, verb); NPN (noun, proper noun); ADJ (adjective)}
}
\value{
list of two data frames - original dataset with topic numbers appended; data of top n terms with topic numbers and PCA coordinates
}
\description{
Complete BTM (topic modeling)
}
\examples{
\dontrun{
complete_btm(data, min_topics, max_topics, "NAV")
}
}
