% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btm_mapping.R
\name{btm_mapping}
\alias{btm_mapping}
\title{Topic Mapping}
\usage{
btm_mapping(models, data, coherence_df, best_topic)
}
\arguments{
\item{models}{List of all potential models run}

\item{data}{The original dataframe, annotated via udpipe}

\item{coherence_df}{Coherence scores for each model}

\item{best_topic}{The selected best topic based upon coherence scores}
}
\description{
Identify the topic for each piece of text in the original dataframe
}
